Convert nose's package setup to a pytest fixture
